Implement optional clear on exit. Fixes #2.
This made it logical to bring back the Monitor class. Monitors
(previously Drivers that did not also inherit Drawable) never
clear the screen on exit because they never own their own buffer.

Instead of Inheriting Driver and Drawable, Driver is now a Monitor
that inherits Drawable and optionally clears the screen at exit.
The default is to clear.